Board: Add place to drop red resin for version
During the assembly phase we want to mark the assembly
number. Problem is if we use a marker it gets washed
away so we can use the red resin in these spots to write
a 8 bit number using dots. Along with a note space for
extra space if we can figure out a non-alchol washable pen.

Signed-off-by: Konrad Rzeszutek Wilk <konrad@kernel.org>