Board/Schematic: Line Sensor: Swap 10K with PIR sensor
I had the 10K with PIR sensor inverted. That is
the voltage would go down when we are in dark room (so
nothing to detect), and go above the Vref (aka the
10K pot on NEG_IN) and light up the LED.

We need it the other way around - that is in dark
rooms (so PIR is 20K ohm ,or 7.8K ohm) we want the
voltage to be below the Vref (NEG_IN), and only go
high when the 940nm light emitter is reflected and
PIR is around 300ohm.

PIR ohm| 10K | voltage|conditions
---+-----+------------+----------
300    | 10K | 4.8V   | light reflected
7.8K   | 10K | 3.6V   | daylight
20K    | 10K | 1.6V   | dark night

Signed-off-by: Konrad Rzeszutek Wilk <konrad@kernel.org>